Store the include path in .htaccess
Write to .htaccess to set a variable. This is the most sensible place
to store the path data. While tempting to actually establish the
includes path in .htaccess, there is no mechanism within .htaccess to
append to the existing includes path. So we'd overwrite any predefined
includes. Instead, we save a variable there, and continue to establish
the includes path within PHP.

Per #370.