Implement spell correction
Suggest a better search term, and provide a link to run that search.
Per #437.

There are two catches, but I'm OK with them. The first is that it uses
string-based replacement. So if somebody searched for "domestic mes"
and Solr suggested "mess" intead of "mes," this would correct the
search term to "domesstic mess." This is an edge case, but it could
happen. The second catch is that the URL for the new search term
ignores any parameters in the URL other than q, such as the number of
results per page.